feat(projectStructure): rename data to dto
