+++ sante.api/src/main/java/org/aksw/sante/api/endpoints/dbpedia/DbpediaLookupOperations.java
@@ -5,7 +5,7 @@ import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.enums.ParameterIn;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import org.aksw.sante.api.endpoints.dbpedia.data.DbpediaDocumentCollection;
+import org.aksw.sante.api.endpoints.dbpedia.dto.DbpediaDocumentCollection;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
