+++ sante.api/src/main/java/org/aksw/sante/api/endpoints/dbpedia/dto/DbpediaDocument.java
@@ -1,4 +1,4 @@
-package org.aksw.sante.api.endpoints.dbpedia.data;
+package org.aksw.sante.api.endpoints.dbpedia.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Getter;
