+++ sante.api/src/main/java/org/aksw/sante/api/endpoints/reconciliation/ReconciliationOperations.java
@@ -6,8 +6,8 @@ import io.swagger.v3.oas.annotations.enums.ParameterIn;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.ExampleObject;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import org.aksw.sante.api.endpoints.reconciliation.data.query.ReconciliationQuery;
-import org.aksw.sante.api.endpoints.reconciliation.data.result.ReconciliationResult;
+import org.aksw.sante.api.endpoints.reconciliation.dto.query.ReconciliationQuery;
+import org.aksw.sante.api.endpoints.reconciliation.dto.result.ReconciliationResult;
 import org.aksw.sante.api.exception.SearchSuggestException;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
