+++ sante.api/src/main/java/org/aksw/sante/api/endpoints/reconciliation/ReconciliationService.java
@@ -3,9 +3,9 @@ package org.aksw.sante.api.endpoints.reconciliation;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.aksw.sante.api.endpoints.reconciliation.data.result.ReconciliationResult;
-import org.aksw.sante.api.endpoints.reconciliation.data.query.ReconciliationQuery;
-import org.aksw.sante.api.endpoints.reconciliation.data.result.ReconciliationCandidate;
+import org.aksw.sante.api.endpoints.reconciliation.dto.result.ReconciliationResult;
+import org.aksw.sante.api.endpoints.reconciliation.dto.query.ReconciliationQuery;
+import org.aksw.sante.api.endpoints.reconciliation.dto.result.ReconciliationCandidate;
 import org.aksw.sante.api.exception.SearchSuggestException;
 import org.aksw.sante.api.wrapper.provider.SearchProvider;
 import org.aksw.sante.entity.Entity;
