+++ sante.api/src/main/java/org/aksw/sante/api/endpoints/santesearch/SanteSearchService.java
@@ -1,7 +1,7 @@
 package org.aksw.sante.api.endpoints.santesearch;
 
-import org.aksw.sante.api.endpoints.santesearch.data.SanteSearchCandidate;
-import org.aksw.sante.api.endpoints.santesearch.data.SanteSearchResponseDto;
+import org.aksw.sante.api.endpoints.santesearch.dto.SanteSearchCandidate;
+import org.aksw.sante.api.endpoints.santesearch.dto.SanteSearchResponseDto;
 import org.aksw.sante.api.exception.SearchSuggestException;
 import org.aksw.sante.api.wrapper.Query;
 import org.aksw.sante.api.wrapper.provider.SearchProvider;
