+++ sante.api/src/main/java/org/aksw/sante/api/endpoints/santesearch/dto/SanteSearchCandidate.java
@@ -1,4 +1,4 @@
-package org.aksw.sante.api.endpoints.santesearch.data;
+package org.aksw.sante.api.endpoints.santesearch.dto;
 
 import lombok.Getter;
 import lombok.Setter;
