+++ sante.api/src/main/java/org/aksw/sante/api/endpoints/santesuggest/SanteSuggestController.java
@@ -1,6 +1,6 @@
 package org.aksw.sante.api.endpoints.santesuggest;
 
-import org.aksw.sante.api.endpoints.santesuggest.data.SanteSuggestResponseDto;
+import org.aksw.sante.api.endpoints.santesuggest.dto.SanteSuggestResponseDto;
 import org.aksw.sante.api.exception.SearchSuggestException;
 import org.aksw.sante.api.wrapper.Query;
 import org.springframework.beans.factory.annotation.Autowired;
